src/init_db/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pathlib
#pymysql
pydantic
kaggle
mysql-connector-python
pandas
scikit-learn
pytest #just to be sure, that the pytests will be executed properly inside the container!